requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
et_xmlfile==2.0.0
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
